refactor: move PageShell to shared components and remove @kiku/core dependency
