--- tanaka/extension/src/playground/components/page-shell/page-shell.scss
@@ -1,30 +1,0 @@
-.tnk-page-shell {
-  background-color: var(--tnk-bg-secondary);
-  --app-shell-padding: var(--tnk-space-md);
-}
-
-.tnk-page-shell__header {
-  background-color: var(--tnk-bg-secondary);
-  border-bottom: var(--tnk-space-none);
-  --app-shell-header-height: var(--tnk-space-7xl);
-}
-
-.tnk-page-shell__brand {
-  height: 100%;
-  padding-left: var(--tnk-space-xl);
-  padding-right: var(--tnk-space-xl);
-}
-
-.tnk-page-shell__brand-text {
-  font-size: var(--tnk-text-4xl);
-  font-weight: var(--tnk-font-weight-bold);
-  background: linear-gradient(135deg, var(--tnk-primary), var(--tnk-primary-dark));
-  background-clip: text;
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  margin: 0;
-}
-
-.tnk-page-shell__main {
-  margin-top: var(--tnk-space-7xl);
-}
